refactor: move shared model to core
