+++ app/src/main/java/com/ecosense/android/core/data/local/dao/SavedRecognitionResultDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.ecosense.android.featRecognition.data.model.SavedRecognitionResultEntity
+import com.ecosense.android.core.data.model.SavedRecognitionResultEntity
 
 @Dao
 interface SavedRecognitionResultDao {
